+++ test/community/training/easy/robotReach/CGTest.js
@@ -1,0 +1,75 @@
+import { assert } from 'chai';
+import sinon from 'sinon';
+import File from '../../../../File.js';
+import { assertOutputAnswer } from '../../../../assertOutputAnswer.js';
+    import { execute } from '../../../../../lib/community/training/easy/robotReach/robotReach.js';
+
+const __dirname = new URL('.', import.meta.url).pathname;
+
+suite("Robot show", function() {
+    const sandbox = sinon.createSandbox();
+
+    setup(function () {
+        sandbox.stub(console, "log");
+    });
+
+    teardown(function () {
+        sandbox.restore();
+    });
+
+
+    test("Simple 3x3", function() {
+        let inputFile = new File(__dirname + 'input/01 - simple 3x3.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            3
+        );
+    });
+
+    test("Simple 3x3 - 3", function() {
+        let inputFile = new File(__dirname + 'input/02 - simple 3x3 - 3.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            8
+        );
+    });
+
+    test("Test 3", function() {
+        let inputFile = new File(__dirname + 'input/03 - test 3.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            59
+        );
+    });
+
+    test("Test 4", function() {
+        let inputFile = new File(__dirname + 'input/04 - test 4.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            636
+        );
+    });
+
+    test("Test 5", function() {
+        let inputFile = new File(__dirname + 'input/05 - test 5.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            3
+        );
+    });
+});