+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Substitution encoding".
 - Tests for "Monday tuesday happy days".
 - Tests for "Magic string".
+- Tests for "Robot reach".
 
 ## [1.9.0] - 2022-06-01
 ### Added